New compilation warnings / fix errors due to implicit conversions
New flags with gcc and clang are:
* -Werror
* -Wshadow
* -Wnon-virtual-dtor
* -Wcast-align
* -Wunused
* -Woverloaded-virtual
* -Wpedantic
* -Wdouble-promotion
* -Wformat=2
* -Wconversion
* -Wsign-conversion

+++ include/highfive/bits/H5Annotate_traits_misc.hpp
@@ -80,7 +80,7 @@ inline size_t AnnotateTraits<Derivate>::getNumberAttributes() const {
         HDF5ErrMapper::ToException<AttributeException>(std::string(
             "Unable to count attributes in existing group or file"));
     }
-    return res;
+    return static_cast<size_t>(res);
 }
 
 template <typename Derivate>
